requirements.txt
Django==2.1.1
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-jsonapi==2.6.0
inflection==0.3.1
pkg-resources==0.0.0
pytz==2018.5
six==1.11.0
